Move sub-classes of "ToolStripDropDownMenu" and "ToolStripPanel" classes to their own files
