+++ src/System.Windows.Forms/src/System/Windows/Forms/ToolStripPanel.cs
@@ -4,20 +4,18 @@
 
 #nullable disable
 
-using System.Collections;
 using System.Collections.Specialized;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Drawing;
 using System.Globalization;
 using System.Windows.Forms.Layout;
-using static Interop;
 
 namespace System.Windows.Forms
 {
     [Designer("System.Windows.Forms.Design.ToolStripPanelDesigner, " + AssemblyRef.SystemDesign)]
     [ToolboxBitmap(typeof(ToolStripPanel), "ToolStripPanel_standalone")]
-    public class ToolStripPanel : ContainerControl, IArrangedElement
+    public partial class ToolStripPanel : ContainerControl, IArrangedElement
     {
         private Orientation _orientation = Orientation.Horizontal;
         private static readonly Padding s_rowMargin = new Padding(3, 0, 0, 0);
@@ -726,168 +724,6 @@ namespace System.Windows.Forms
             }
         }
 
-        // The FeedbackRectangle happens to encapsulate a toolstripdropdown
-        // with a special region. The feedback rectangle exposes the minimum
-        // API so the underlying implementation can be replaced if necessary.
-        private class FeedbackRectangle : IDisposable
-        {
-            private FeedbackDropDown _dropDown;
-
-            public FeedbackRectangle(Rectangle bounds)
-            {
-                _dropDown = new FeedbackDropDown(bounds);
-            }
-
-            public bool Visible
-            {
-                get
-                {
-                    if (_dropDown != null && !_dropDown.IsDisposed)
-                    {
-                        return _dropDown.Visible;
-                    }
-                    return false;
-                }
-                set
-                {
-                    if (_dropDown != null && !_dropDown.IsDisposed)
-                    {
-                        _dropDown.Visible = value;
-                    }
-                }
-            }
-
-            public void Show(Point newLocation)
-            {
-                _dropDown.Show(newLocation);
-            }
-            public void Move(Point newLocation)
-            {
-                _dropDown.MoveTo(newLocation);
-            }
-
-            protected void Dispose(bool disposing)
-            {
-                if (disposing)
-                {
-                    if (_dropDown != null)
-                    {
-                        Visible = false;
-                        _dropDown.Dispose();
-                        _dropDown = null;
-                    }
-                }
-            }
-
-            public void Dispose()
-            {
-                Dispose(true);
-            }
-
-            ~FeedbackRectangle()
-            {
-                Dispose(false);
-            }
-
-            private class FeedbackDropDown : ToolStripDropDown
-            {
-                private const int MaxPaintsToService = 20;
-                private int _numPaintsServiced; // member variable to protect against re-entrancy
-
-                public FeedbackDropDown(Rectangle bounds) : base()
-                {
-                    SetStyle(ControlStyles.AllPaintingInWmPaint, false);
-                    SetStyle(ControlStyles.OptimizedDoubleBuffer, true);
-                    SetStyle(ControlStyles.CacheText, true);
-                    AutoClose = false;
-                    AutoSize = false;
-                    DropShadowEnabled = false;
-                    Bounds = bounds;
-
-                    // caching as this is unlikely to change during the lifetime
-                    // of the dropdown.
-
-                    Rectangle regionRect = bounds;    //create a region the size of the client area
-                    regionRect.Inflate(-1, -1);        //squish down by one pixel
-
-                    Region rgn = new Region(bounds);  // create region
-                    rgn.Exclude(regionRect);          // exclude the center part
-
-                    // set it into the toolstripdropdown’s region
-                    Region = rgn;
-                }
-
-                // ForceSynchronousPaint - peeks through the message queue, looking for WM_PAINTs
-                // calls UpdateWindow on the hwnd to force the paint to happen now.
-                //
-                // When we're changing the location of the feedback dropdown, we need to
-                // force WM_PAINTS to happen, as things that don't respond to WM_ERASEBKGND
-                // have bits of the dropdown region drawn all over them.
-                private void ForceSynchronousPaint()
-                {
-                    if (!IsDisposed)
-                    {
-                        if (_numPaintsServiced == 0)
-                        {
-                            // protect against re-entrancy.
-                            try
-                            {
-                                var msg = new User32.MSG();
-                                while (User32.PeekMessageW(ref msg, IntPtr.Zero, User32.WM.PAINT, User32.WM.PAINT, User32.PM.REMOVE).IsTrue())
-                                {
-                                    User32.UpdateWindow(msg.hwnd);
-
-                                    // Infinite loop protection
-                                    if (_numPaintsServiced++ > MaxPaintsToService)
-                                    {
-                                        Debug.Fail("somehow we've gotten ourself in a situation where we're pumping an unreasonable number of paint messages, investigate.");
-                                        break;
-                                    }
-                                }
-                            }
-                            finally
-                            {
-                                _numPaintsServiced = 0;
-                            }
-                        }
-                    }
-                }
-
-                protected override void OnPaint(PaintEventArgs e)
-                {
-                }
-                protected override void OnPaintBackground(PaintEventArgs e)
-                {
-                    // respond to everything in WM_ERASEBKGND
-                    Renderer.DrawToolStripBackground(new ToolStripRenderEventArgs(e.Graphics, this));
-                    Renderer.DrawToolStripBorder(new ToolStripRenderEventArgs(e.Graphics, this));
-                }
-                protected override void OnOpening(CancelEventArgs e)
-                {
-                    base.OnOpening(e);
-                    e.Cancel = false;
-                }
-
-                public void MoveTo(Point newLocation)
-                {
-                    Location = newLocation;
-                    // if we don't force a paint here, we'll only send WM_ERASEBKGNDs right away
-                    // and leave rectangles all over controls that don't respond to that window message.
-                    ForceSynchronousPaint();
-                }
-
-                protected override void WndProc(ref Message m)
-                {
-                    if (m.Msg == (int)User32.WM.NCHITTEST)
-                    {
-                        m.Result = (IntPtr)User32.HT.TRANSPARENT;
-                    }
-
-                    base.WndProc(ref m);
-                }
-            }
-        }
-
         #endregion Feedback
 
         #region JoinAndMove
@@ -1339,299 +1175,6 @@ namespace System.Windows.Forms
             get
             {
                 return RowsInternal;
-            }
-        }
-
-        [ListBindable(false)]
-        public class ToolStripPanelRowCollection : ArrangedElementCollection, IList
-        {
-            private readonly ToolStripPanel _owner;
-
-            public ToolStripPanelRowCollection(ToolStripPanel owner)
-            {
-                _owner = owner;
-            }
-
-            public ToolStripPanelRowCollection(ToolStripPanel owner, ToolStripPanelRow[] value)
-            {
-                _owner = owner;
-                AddRange(value);
-            }
-
-            /// <summary>
-            ///
-            /// </summary>
-            public new virtual ToolStripPanelRow this[int index]
-            {
-                get
-                {
-                    return (ToolStripPanelRow)(InnerList[index]);
-                }
-            }
-
-            public int Add(ToolStripPanelRow value)
-            {
-                if (value is null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-                int retVal = InnerList.Add(value);
-                OnAdd(value, retVal);
-                return retVal;
-            }
-
-            public void AddRange(ToolStripPanelRow[] value)
-            {
-                if (value is null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                ToolStripPanel currentOwner = _owner;
-                if (currentOwner != null)
-                {
-                    currentOwner.SuspendLayout();
-                }
-                try
-                {
-                    for (int i = 0; i < value.Length; i++)
-                    {
-                        Add(value[i]);
-                    }
-                }
-                finally
-                {
-                    if (currentOwner != null)
-                    {
-                        currentOwner.ResumeLayout();
-                    }
-                }
-            }
-            public void AddRange(ToolStripPanelRowCollection value)
-            {
-                if (value is null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                ToolStripPanel currentOwner = _owner;
-                if (currentOwner != null)
-                {
-                    currentOwner.SuspendLayout();
-                }
-                try
-                {
-                    int currentCount = value.Count;
-                    for (int i = 0; i < currentCount; i++)
-                    {
-                        Add(value[i]);
-                    }
-                }
-                finally
-                {
-                    if (currentOwner != null)
-                    {
-                        currentOwner.ResumeLayout();
-                    }
-                }
-            }
-
-            public bool Contains(ToolStripPanelRow value)
-            {
-                return InnerList.Contains(value);
-            }
-
-            public virtual void Clear()
-            {
-                if (_owner != null)
-                {
-                    _owner.SuspendLayout();
-                }
-                try
-                {
-                    while (Count != 0)
-                    {
-                        RemoveAt(Count - 1);
-                    }
-                }
-                finally
-                {
-                    if (_owner != null)
-                    {
-                        _owner.ResumeLayout();
-                    }
-                }
-            }
-
-            void IList.Clear() { Clear(); }
-            bool IList.IsFixedSize { get { return InnerList.IsFixedSize; } }
-            bool IList.Contains(object value) { return InnerList.Contains(value); }
-            bool IList.IsReadOnly { get { return InnerList.IsReadOnly; } }
-            void IList.RemoveAt(int index) { RemoveAt(index); }
-            void IList.Remove(object value) { Remove(value as ToolStripPanelRow); }
-            int IList.Add(object value) { return Add(value as ToolStripPanelRow); }
-            int IList.IndexOf(object value) { return IndexOf(value as ToolStripPanelRow); }
-            void IList.Insert(int index, object value) { Insert(index, value as ToolStripPanelRow); }
-
-            object IList.this[int index]
-            {
-                get { return InnerList[index]; }
-                set { throw new NotSupportedException(SR.ToolStripCollectionMustInsertAndRemove); /* InnerList[index] = value; */ }
-            }
-
-            public int IndexOf(ToolStripPanelRow value)
-            {
-                return InnerList.IndexOf(value);
-            }
-
-            public void Insert(int index, ToolStripPanelRow value)
-            {
-                if (value is null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                InnerList.Insert(index, value);
-                OnAdd(value, index);
-            }
-
-            private void OnAdd(ToolStripPanelRow value, int index)
-            {
-                if (_owner != null)
-                {
-                    LayoutTransaction.DoLayout(_owner, value, PropertyNames.Parent);
-                }
-            }
-
-            /// <summary>
-            ///  Do proper cleanup of ownership, etc.
-            /// </summary>
-            private void OnAfterRemove(ToolStripPanelRow row)
-            {
-#if DEBUG
-                if (s_toolStripPanelMissingRowDebug.TraceVerbose)
-                {
-                    if (row != null)
-                    {
-                        Debug.Write("Removing row: ");
-                        row.Debug_PrintRowID();
-                        Debug.WriteLine(new StackTrace().ToString());
-                    }
-                }
-#endif
-
-            }
-
-            public void Remove(ToolStripPanelRow value)
-            {
-                InnerList.Remove(value);
-                OnAfterRemove(value);
-            }
-
-            public void RemoveAt(int index)
-            {
-                ToolStripPanelRow item = null;
-                if (index < Count && index >= 0)
-                {
-                    item = (ToolStripPanelRow)(InnerList[index]);
-                }
-                InnerList.RemoveAt(index);
-                OnAfterRemove(item);
-            }
-
-            public void CopyTo(ToolStripPanelRow[] array, int index)
-            {
-                InnerList.CopyTo(array, index);
-            }
-        }
-
-        internal class ToolStripPanelControlCollection : TypedControlCollection
-        {
-            private readonly ToolStripPanel _owner;
-
-            public ToolStripPanelControlCollection(ToolStripPanel owner)
-                : base(owner, typeof(ToolStrip))
-            {
-                _owner = owner;
-            }
-
-            internal override void AddInternal(Control value)
-            {
-                if (value != null)
-                {
-                    using (new LayoutTransaction(value, value, PropertyNames.Parent))
-                    {
-                        base.AddInternal(value);
-                    }
-                }
-                else
-                {
-                    base.AddInternal(value);
-                }
-            }
-
-            internal void Sort()
-            {
-                if (_owner.Orientation == Orientation.Horizontal)
-                {
-                    InnerList.Sort(new YXComparer());
-                }
-                else
-                {
-                    InnerList.Sort(new XYComparer());
-                }
-            }
-
-            // sort by X, then Y
-            public class XYComparer : IComparer
-            {
-                public XYComparer() { }
-                public int Compare(object first, object second)
-                {
-                    Control one = first as Control;
-                    Control two = second as Control;
-
-                    if (one.Bounds.X < two.Bounds.X)
-                    {
-                        return -1;
-                    }
-
-                    if (one.Bounds.X == two.Bounds.X)
-                    {
-                        if (one.Bounds.Y < two.Bounds.Y)
-                        {
-                            return -1;
-                        }
-                        return 1;
-                    }
-                    return 1;
-                }
-            }
-
-            // sort by Y, then X
-            public class YXComparer : IComparer
-            {
-                public YXComparer() { }
-                public int Compare(object first, object second)
-                {
-                    Control one = first as Control;
-                    Control two = second as Control;
-
-                    if (one.Bounds.Y < two.Bounds.Y)
-                    {
-                        return -1;
-                    }
-
-                    if (one.Bounds.Y == two.Bounds.Y)
-                    {
-                        if (one.Bounds.X < two.Bounds.X)
-                        {
-                            return -1;
-                        }
-                        return 1;
-                    }
-                    return 1;
-                }
             }
         }
     }
